+++ HISTORY.rst
@@ -57,6 +57,8 @@ Unreleased Changes
 * Habitat Quality
     * Access raster is now generated from the reprojected access vector
       (https://github.com/natcap/invest/issues/1615).
+    * Rarity values are now output in CSV format (as well as in raster format)
+      (https://github.com/natcap/invest/issues/721).
 * Urban Flood Risk
     * Fields present on the input AOI vector are now retained in the output.
       (https://github.com/natcap/invest/issues/1600)